refactor-and-fix-openaichat | lint
