+++ embeddings/openai/openaichat/openai_chat_test.go
@@ -1,4 +1,4 @@
-package openai_chat
+package openaichat
 
 import (
 	"context"
@@ -7,7 +7,6 @@ import (
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-
 	"github.com/tmc/langchaingo/llms/openai"
 )
